+++ ShapeCrawler/Exceptions/NotSupportedFeature.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace ShapeCrawler.Exceptions
+﻿namespace ShapeCrawler.Exceptions
 {
     /// <summary>
     ///     Thrown when the feature has not yet been supported.
@@ -9,15 +7,6 @@ namespace ShapeCrawler.Exceptions
     {
         public NotSupportedFeature(string message)
             : base(message, ExceptionCode.NextVersionFeatureException)
-        {
-        }
-
-        public NotSupportedFeature()
-        {
-        }
-
-        public NotSupportedFeature(string message, Exception innerException)
-            : base(message, innerException)
         {
         }
     }