+++ ShapeCrawler/Exceptions/SlideMasterPropertyCannotBeChanged.cs
@@ -1,19 +1,9 @@
-﻿using System;
-
-namespace ShapeCrawler.Exceptions
+﻿namespace ShapeCrawler.Exceptions
 {
     internal class SlideMasterPropertyCannotBeChanged : ShapeCrawlerException
     {
-        public SlideMasterPropertyCannotBeChanged()
-        {
-        }
-
-        public SlideMasterPropertyCannotBeChanged(string message) : base(message)
-        {
-        }
-
-        public SlideMasterPropertyCannotBeChanged(string message, Exception innerException) : base(message,
-            innerException)
+        internal SlideMasterPropertyCannotBeChanged(string message)
+            : base(message)
         {
         }
     }