+++ ShapeCrawler/PowerPoint/IPresentation.cs
@@ -35,6 +35,9 @@ namespace ShapeCrawler
         /// </summary>
         byte[] ByteArray { get; }
 
+        /// <summary>
+        ///     Gets section collection.
+        /// </summary>
         ISectionCollection Sections { get; }
 
         /// <summary>
@@ -45,7 +48,7 @@ namespace ShapeCrawler
         /// <summary>
         ///     Saves presentation in specified file path.
         /// </summary>
-        void SaveAs(string filePath);
+        void SaveAs(string path);
 
         /// <summary>
         ///     Saves presentation in specified stream.